Documents catalog refactor
