+++ src/main/java/pl/com/bottega/documentmanagement/api/DocumentsCatalog.java
@@ -5,7 +5,6 @@ import pl.com.bottega.documentmanagement.domain.Document;
 import pl.com.bottega.documentmanagement.domain.DocumentCriteria;
 import pl.com.bottega.documentmanagement.domain.DocumentNumber;
 import pl.com.bottega.documentmanagement.domain.repositories.DocumentRepository;
-import pl.com.bottega.documentmanagement.infrastructure.DocumentDto;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -14,29 +13,16 @@ import static com.google.common.base.Preconditions.checkNotNull;
  */
 public class DocumentsCatalog {
 
-    private DocumentRepository documentRepository;
-
     public DocumentDto get(DocumentNumber documentNumber) {
         checkNotNull(documentNumber);
-        Document document = documentRepository.load(documentNumber);
 
-        return document.export();
+        return null;
     }
 
     public Iterable<DocumentDto> find(DocumentCriteria documentCriteria) {
         checkNotNull(documentCriteria);
-        Iterable<Document> documents = documentRepository.find(documentCriteria);
 
-        // Równoważe zapisy:
-
-        /*Collection<DocumentDto> returnValue = new ArrayList<>();
-        for(Document document : documents)
-            returnValue.add(document.export());
-        return returnValue;*/
-
-        //return Iterables.transform(documents, (document) -> document.export());
-
-        return Iterables.transform(documents, Document::export);
+        return null;
     }
 
 }
